requirements.txt
beautifulsoup4==4.14.3
PyGithub==2.8.1
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8.3
tqdm==4.67.3
numpy==2.0.2
pandas==2.3.3
